Prevent to go into auth pages when already logged in
