+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { ReactNode } from "react";
-import Link from "next/link";
 
 import NextTopLoader from "nextjs-toploader";
-import Navbar from "@/components/Navbar";
+import NavbarWrapper from "@/components/NavbarWrapper";
+
 import Footer from "@/components/Footer";
 
 import "../styles/globals.scss";
@@ -14,7 +14,7 @@ export default function RootLayout({ children }: { children: ReactNode }) {
 				<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" rel="stylesheet" />
 			</head>
 			<body style={{ fontFamily: "Inter, Arial, sans-serif" }}>
-				<Navbar />
+				<NavbarWrapper />
 				<NextTopLoader showSpinner={false} height={10} />
 				{children}
 				<Footer />
