+++ components/NavbarWrapper.tsx
@@ -1,0 +1,10 @@
+// components/NavbarWrapper.tsx
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
+import Navbar from "./Navbar";
+
+export default async function NavbarWrapper() {
+	const session = await getServerSession(authOptions);
+
+	return <Navbar session={session} />;
+}
